feat: Improve UX when user creates or deletes post (PostForm/PostItem) - Route to MainPage when user creates post in PostForm component - Confirm when user deletes post in PostItem component

+++ vue-til/src/components/posts/PostItem.vue
@@ -24,8 +24,10 @@ export default {
 
   methods: {
     async deleteItem() {
-      await deletePost(this.item._id);
-      this.$emit('refresh');
+      if (confirm('Are you sure you want to delete?')) {
+        await deletePost(this.item._id);
+        this.$emit('refresh');
+      }
     },
   },
 };
